Reorganize, fixed debug
