+++ extract_msg/__doc__.py
@@ -1,0 +1,6 @@
+"""
+extract_msg:
+    Extracts emails and attachments saved in Microsoft Outlook's .msg files
+
+https://github.com/mattgwwalker/msg-extractor
+"""
