+++ extract_msg/constants.py
@@ -1,5 +1,7 @@
 import struct
 
+
+
 # DEFINE CONSTANTS
 # WARNING DO NOT CHANGE ANY OF THESE VALUES UNLESS YOU KNOW
 # WHAT YOU ARE DOING! FAILURE TO FOLLOW THIS INSTRUCTION
