+++ extract_msg/debug.py
@@ -1,0 +1,1 @@
+debug = False
